refactor: float & integer in own modules
